refactor: changed to package & base queue config add
