+++ sample/src/main/java/com/example/messagingrabbitmq/base/CustomMessage.java
@@ -1,4 +1,4 @@
-package com.example.messagingrabbitmq.code;
+package com.example.messagingrabbitmq.base;
 
 public record CustomMessage(String text, int index) {
 }
