+++ sample/src/main/java/com/example/messagingrabbitmq/base/SampleListener.java
@@ -1,4 +1,4 @@
-package com.example.messagingrabbitmq.code;
+package com.example.messagingrabbitmq.base;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class SampleListener {
 
-    @RabbitListener(queues = RabbitMaConfiguration.queueName)
+    @RabbitListener(queues = BaseRabbitMqInfo.queueName)
     public void receiveMessage(CustomMessage message) {
         log.info("get message : {}", message);
     }
