+++ sample/src/main/java/com/example/messagingrabbitmq/config/RabbitMqConfiguration.java
@@ -1,0 +1,40 @@
+package com.example.messagingrabbitmq.config;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.amqp.rabbit.connection.CachingConnectionFactory;
+import org.springframework.amqp.rabbit.connection.ConnectionFactory;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
+import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
+import org.springframework.amqp.support.converter.MessageConverter;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@RequiredArgsConstructor
+public class RabbitMqConfiguration {
+
+    private final RabbitMqProperties rabbitMqProperties;
+
+    @Bean
+    public ConnectionFactory connectionFactory() {
+        CachingConnectionFactory connectionFactory = new CachingConnectionFactory();
+        connectionFactory.setHost(rabbitMqProperties.getHost());
+        connectionFactory.setPort(rabbitMqProperties.getPort());
+        connectionFactory.setUsername(rabbitMqProperties.getUsername());
+        connectionFactory.setPassword(rabbitMqProperties.getPassword());
+        return connectionFactory;
+    }
+
+    @Bean
+    public MessageConverter messageConverter() {
+        return new Jackson2JsonMessageConverter();
+    }
+
+    @Bean
+    public RabbitTemplate rabbitTemplate(ConnectionFactory connectionFactory, MessageConverter messageConverter) {
+        RabbitTemplate rabbitTemplate = new RabbitTemplate(connectionFactory);
+        rabbitTemplate.setMessageConverter(messageConverter);
+        return rabbitTemplate;
+    }
+
+}
